Add script to fit models (to be simplified)
This script is copy-pasted from a previous one that contained all analyses and plots, including the ones for the norming study. In this dataverse repos I will split some of these analyses into different files to make things clearer. This means much of this script will be cut out or otherwise simplified.